+++ src/wtforms/form.py
@@ -278,7 +278,11 @@ class Form(BaseForm, metaclass=FormMeta):
         meta_obj = self._wtforms_meta()
         if meta is not None and isinstance(meta, dict):
             meta_obj.update_values(meta)
-        super().__init__(self._unbound_fields, meta=meta_obj, prefix=prefix)
+        super().__init__(
+            self._unbound_fields,
+            meta=meta_obj,
+            prefix=prefix,
+        )
 
         for name, field in self._fields.items():
             # Set all the fields to attributes so that they obscure the class